requirements.txt
django
djangorestframework
drf_spectacular
django-filter
django-simple-history
psycopg2
gunicorn
pytest